1. Removed the array of skill properties and replaced them with explicit field variables. 2. Removed ID from objects with templates as it was redundant to store the ID inside the object and it's associated template. 3. Stopped using uints, it got annoying type casting. 4. Finished Field loading. 5. First implementation of Mob and TemplateMob.

+++ Common/Game/SpawnPoint.cs
@@ -16,7 +16,7 @@ namespace NineToFive.Game {
             if (Mob != null) return;
 
             Mob = new Mob(MobID);
-            Field.Life[EntityType.Mob].Add((uint)MobID, Mob);
+            Field.Life[EntityType.Mob].Add(MobID, Mob);
             
             //todo: spawn
         }
